requirements.txt
astroid==2.12.12
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dill==0.3.6
elastic-transport==8.4.0
elasticsearch==8.4.3
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
flask-swagger-ui==4.11.1
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
nltk==3.7
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
platformdirs==2.5.2
pycodestyle==2.9.1
pylint==2.15.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
threadpoolctl==3.1.0
tokenizers==0.13.1
tomli==2.0.1
tomlkit==0.11.6
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.23.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
